+++ packages/image_picker/image_picker/example/ios/RunnerTests/PhotoAssetUtilTests.m
@@ -101,11 +101,10 @@
 
   size_t numberOfFrames = CGImageSourceGetCount(imageSource);
 
-  NSNumber *nilSize = (NSNumber *)[NSNull null];
   NSString *savedPathGIF = [FLTImagePickerPhotoAssetUtil saveImageWithOriginalImageData:dataGIF
                                                                                   image:imageGIF
-                                                                               maxWidth:nilSize
-                                                                              maxHeight:nilSize
+                                                                               maxWidth:nil
+                                                                              maxHeight:nil
                                                                            imageQuality:nil];
   XCTAssertNotNil(savedPathGIF);
   XCTAssertEqualObjects([savedPathGIF substringFromIndex:savedPathGIF.length - 4], @".gif");